requirements.txt
Flask>=2.0
pycryptodome
gunicorn
